@@ -19,7 +19,7 @@ struct EntityRecord {
 
 /// Types for defining Archetypes, collections of entities that have the same set of
 /// components.
-public final class Archetype: Hashable, Identifiable, Sendable {
+public final class Archetype: Hashable, Identifiable, @unchecked Sendable {
     public let id: Int
     public internal(set) var entities: SparseArray<Entity> = []
     
