@@ -8,7 +8,7 @@
 import Foundation
 
 /// An object that manages image data in your app.
-public final class Image {
+public final class Image: @unchecked Sendable {
     
     public private(set) var data: Data
     
