@@ -8,7 +8,7 @@
 import Foundation
 
 /// An interface that describe how to build an ``Image`` object from bytes.
-protocol ImageLoaderStrategy {
+protocol ImageLoaderStrategy: Sendable {
     
     func canDecodeImage(with fileExtensions: String) -> Bool
     
