@@ -14,47 +14,63 @@
 /// If you want serialize your ``Component`` object, annotate properties inside them as `@Export`.
 /// - Note: You can use `private`, `fileprivate` modifiers, because `@Export` use reflection
 @propertyWrapper
-public final class Export<T: Codable>: Codable, @unchecked Sendable {
-    
+public struct Export<T: Codable>: Codable, @unchecked Sendable {
+
+    private final class Storage {
+        var skipped: Bool = false
+        var hasChanges: Bool = false
+        var editorInfo: EditorInfo?
+        var value: T {
+            didSet {
+                self.hasChanges = true
+            }
+        }
+
+        init(value: T) {
+            self.value = value
+        }
+    }
+
+    private let storage: Storage
+
     public var wrappedValue: T {
-        didSet {
-            self.hasChanges = true
+        get {
+            self.storage.value
+        }
+        set {
+            self.storage.value = newValue
         }
     }
     
-    private var skipped: Bool = false
-    private var hasChanges: Bool = false
-    private var editorInfo: EditorInfo?
-    
     public init(wrappedValue: T) {
-        self.wrappedValue = wrappedValue
+        self.storage = .init(value: wrappedValue)
     }
     
     public init(wrappedValue: T) where T: Resource {
-        self.wrappedValue = wrappedValue
-        self.editorInfo = EditorInfo(modifiers: .resource)
+        self.storage = .init(value: wrappedValue)
+        self.storage.editorInfo = EditorInfo(modifiers: .resource)
     }
     
     public init(wrappedValue: T) where T: CaseIterable {
-        self.wrappedValue = wrappedValue
-        self.editorInfo = EditorInfo(
+        self.storage = .init(value: wrappedValue)
+        self.storage.editorInfo = EditorInfo(
             modifiers: .enum(EnumModifier(cases: T.allCases.map { String(describing: $0) }))
         )
     }
     
     // MARK: - Codable
     
-    public required init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
-        self.wrappedValue = try container.decode(T.self)
+        self.storage = .init(value: try container.decode(T.self))
     }
     
     public func encode(to encoder: Encoder) throws {
-        if encoder.userInfo[.editorIntrospection] != nil && self.editorInfo != nil {
+        if encoder.userInfo[.editorIntrospection] != nil && self.storage.editorInfo != nil {
             var container = encoder.container(keyedBy: CodingName.self)
             try container.encode(self.wrappedValue, forKey: .value)
-            try container.encode(self.editorInfo, forKey: .editor)
-        } else if self.hasChanges {
+            try container.encode(self.storage.editorInfo, forKey: .editor)
+        } else if self.storage.hasChanges {
             var container = encoder.singleValueContainer()
             try container.encode(self.wrappedValue)
         }
@@ -69,32 +85,32 @@ extension Export: _ExportCodable {
             return
         }
         
-        self.wrappedValue = value
+        self.storage.value = value
     }
-     
+    
     func encode(to container: inout EncodingContainer, propertyName: String, userInfo: [CodingUserInfoKey: Any]) throws {
         /// we have difference between initial value
-        if self.hasChanges && !self.skipped {
+        if self.storage.hasChanges && !self.storage.skipped {
             try container.encode(self.wrappedValue, forKey: CodingName(stringValue: propertyName))
         }
         
         if userInfo[.editorIntrospection] != nil {
-            try container.encodeIfPresent(self.editorInfo, forKey: .editor)
+            try container.encodeIfPresent(self.storage.editorInfo, forKey: .editor)
         }
     }
 }
 
 public extension Export {
-    convenience init(wrappedValue: T, range: ClosedRange<Float>? = nil, stride: Float? = nil) where T: FloatingPoint {
+    init(wrappedValue: T, range: ClosedRange<Float>? = nil, stride: Float? = nil) where T: FloatingPoint {
         self.init(wrappedValue: wrappedValue)
-        self.editorInfo = EditorInfo(
+        self.storage.editorInfo = EditorInfo(
             modifiers: .float(FloatingPointModifier(range: range, stride: stride))
         )
     }
     
-    convenience init(wrappedValue: T, skipped: Bool) {
+    init(wrappedValue: T, skipped: Bool) {
         self.init(wrappedValue: wrappedValue)
-        self.skipped = skipped
+        self.storage.skipped = skipped
     }
 }
 
